# Conflicts:
#	field-node/files/move_old_subdirs.sh

+++ field-node/files/move_old_subdirs.sh
@@ -11,7 +11,7 @@ if [ "$#" -ne 3 ] || ! [ -d "$1" ] || ! [ -d "$2" ]; then
   echo ""
   echo "Where:" 
   echo "        dir_to_search is the source directory (directory to look within)"
-  echo "        dir_to_move_to is the destination directory (directory to move old directories to)"
+  echo "        dir_to_move_to is the destination mountpoint (directory to move old directories to)"
   echo "        age is the older-than time (in days); directories older than this will be moved"
   exit 1
 fi
@@ -21,9 +21,11 @@ DEST_DIR="$2"
 MOVE_OLDER_THAN_DAYS="$3"
 
 if [ -d "$DEST_DIR" ]; then
-    #  - On OSX we can use -Btime to use the inode birth time, but on Linux mtime is the best we can check
-    #  - Note the use of the wildcard in "$SOURCE_DIR/*" and "-maxdepth 0", and "-d": this limits the search
-    #    to only directories within the source dir (but not the source dir itself)
-    find $SOURCE_DIR/* -maxdepth 0 -type d -mtime +$MOVE_OLDER_THAN_DAYS -exec cp -R --no-preserve=mode,ownership "{}" $DEST_DIR \; -exec rm -r "{}" \;
+    # if the destination is a mounted drive, copy to it
+    if $(mount | grep "$DEST_DIR" &> /dev/null); then
+        #  - On OSX we can use -Btime to use the inode birth time, but on Linux mtime is the best we can check
+        #  - Note the use of the wildcard in "$SOURCE_DIR/*" and "-maxdepth 0", and "-d": this limits the search
+        #    to only directories within the source dir (but not the source dir itself)
+        find $SOURCE_DIR/* -maxdepth 0 -type d -mtime +$MOVE_OLDER_THAN_DAYS -exec cp -R --no-preserve=mode,ownership "{}" $DEST_DIR \; -exec rm -r "{}" \;
+    fi
 fi
-
